First batch of Odyssey changes to existing events.
ShipIDs must now be long, it appears that 'on-foot' has a high ID number, beyond Int32.
Added new ranks to Rank/Progress/Promotion. (Confirmed Progress values are int.)
Added optional Taxi/Multicrew properties to Location/docking/jump events - these are currently undocumented, and not clear if optional for Horizons.

+++ src/Journal/Howatworks.SubEtha.Journal/StationServices/MassModuleStore.cs
@@ -25,7 +25,7 @@ namespace Howatworks.SubEtha.Journal.StationServices
         public long MarketID { get; set; }
         public string Ship { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int ShipId { get; set; } // TODO: check name
+        public long ShipId { get; set; } // TODO: check name
         public List<ModuleItem> Items { get; set; }
     }
