+++ src/Journal/Howatworks.SubEtha.Journal/StationServices/ModuleBuy.cs
@@ -13,7 +13,7 @@ namespace Howatworks.SubEtha.Journal.StationServices
         public long BuyPrice { get; set; }
         public string Ship { get; set; } // Note: ship type
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int ShipID { get; set; }
+        public long ShipID { get; set; }
 
         public string StoredItem { get; set; }
