+++ src/Journal/Howatworks.SubEtha.Journal/StationServices/ModuleRetrieve.cs
@@ -10,7 +10,7 @@ namespace Howatworks.SubEtha.Journal.StationServices
         public string Slot { get; set; }
         public string Ship { get; set; } // Note: ship type
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int ShipID { get; set; }
+        public long ShipID { get; set; }
         public string RetrievedItem { get; set; }
         public string RetrievedItem_Localised { get; set; }
         public bool Hot { get; set; }
