+++ src/Journal/Howatworks.SubEtha.Journal/StationServices/SellShipOnRebuy.cs
@@ -8,7 +8,7 @@ namespace Howatworks.SubEtha.Journal.StationServices
         public string ShipType { get; set; }
         public string System { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int SellShipID { get; set; } // TODO: Check capitalisation
+        public long SellShipID { get; set; } // TODO: Check capitalisation
         public long ShipPrice { get; set; }
     }
 }
