+++ src/Journal/Howatworks.SubEtha.Journal/StationServices/ShipyardBuy.cs
@@ -11,10 +11,10 @@ namespace Howatworks.SubEtha.Journal.StationServices
         public long ShipPrice { get; set; }
         public string StoreOldShip { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int StoreShipID { get; set; }
+        public long StoreShipID { get; set; }
         public string SellOldShip { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int SellShipID { get; set; }
+        public long SellShipID { get; set; }
         public long SellPrice { get; set; }
     }
 }
