+++ src/Journal/Howatworks.SubEtha.Journal/StationServices/ShipyardSwap.cs
@@ -10,12 +10,12 @@ namespace Howatworks.SubEtha.Journal.StationServices
         public string ShipType { get; set; }
         public string ShipType_Localised { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int ShipID { get; set; }
+        public long ShipID { get; set; }
         public string StoreOldShip { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int StoreShipID { get; set; }
+        public long StoreShipID { get; set; }
         public string SellOldShip { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
-        public int SellShipID { get; set; }
+        public long SellShipID { get; set; }
     }
 }
